Move all inline JavaScript for templates/covers to single file. Fix button for "Finished" under templates/saved
