+++ openlibrary/templates/covers/saved.html
@@ -7,7 +7,7 @@ $putctx("cssfile", "form")
         <img src="$image.url(size='M')" class="cover" alt="New book cover"/>
 </div>
 
-<div class="imageSaved", data-image-id="$(image and image.id)">
+<div class="imageSaved" data-image-id="$(image and image.id)">
     <div class="saved">$_("Saved!")</div>
     <div class="meta">
         $if showinfo:
